requirements.txt
requests==2.31.0
alpaca-trade-api==3.1.1
python-telegram-bot==20.7
apscheduler==3.10.4
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.40
arch==6.3.0
scikit-learn==1.4.2
python-dotenv==1.0.1
oandapyV20==0.7.2
krakenex==2.1.0
